Add JaxRs resource scanner class.
This class is used to scan a project for classes that can
act as JaxRs resources.  A class is identified as a JaxRs
resource when it is decorated with the @Path() decorator.

The resource scanner can be given a single string or list of
strings at construction time, each representing either a
fully-qualified class name or a package to scan.  A single
string with multiple resources, comma-separated, can also
be provided.  It can perhaps most easily be used by configuring
it via dependency injection.  A single string parameter
representing one or more class or package names, comma separated,
can be configured via the injector allowing the scanning to be
fully configurable.