Add EagerCachingJaxRsResourceMapper class.
This class is an implementation of JaxRsResourceMapper with
two specific features:

1. It uses a LoadingCache to keep track of requests that have
previously been matched to a resource so the correct handler
doesn't have to be computed again.
2. It is eager in the sense that it preloads all known resource
endpoints into memory at instantiation time - it doesn't have to
scan the resource classes or packages at request time to look for
a match (as compared to a hypothetical non-eager mapper, which
would scan when a request is received).

Because of its characteristics this mapper should be slightly
slower at startup time but give better performance, especially
once the cache has been prepopulated.  Drawbacks include the
inability to respond dynamically to changes to resource
identification, both because of the caching and the fact that
resources are pre-loaded.